fix(boundaries): move desktop runtime detection to config layer
