+++ src/config/runtime.ts
@@ -1,0 +1,32 @@
+export type RuntimeProbe = {
+  hasTauriGlobals: boolean;
+  userAgent: string;
+  locationProtocol: string;
+  locationHost: string;
+  locationOrigin: string;
+};
+
+export function detectDesktopRuntime(probe: RuntimeProbe): boolean {
+  const tauriInUserAgent = probe.userAgent.includes('Tauri');
+  const secureLocalhostOrigin = (
+    probe.locationProtocol === 'https:' && (
+      probe.locationHost === 'localhost' ||
+      probe.locationHost.startsWith('localhost:') ||
+      probe.locationHost === '127.0.0.1' ||
+      probe.locationHost.startsWith('127.0.0.1:')
+    )
+  );
+
+  // Tauri production windows can expose tauri-like hosts/schemes without
+  // always exposing bridge globals at first paint.
+  const tauriLikeLocation = (
+    probe.locationProtocol === 'tauri:' ||
+    probe.locationProtocol === 'asset:' ||
+    probe.locationHost === 'tauri.localhost' ||
+    probe.locationHost.endsWith('.tauri.localhost') ||
+    probe.locationOrigin.startsWith('tauri://') ||
+    secureLocalhostOrigin
+  );
+
+  return probe.hasTauriGlobals || tauriInUserAgent || tauriLikeLocation;
+}
